requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.55
langchain-community>=0.3.9
langchain-openai>=0.3.7
langchain-core>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file reading

# Async support
asyncio

# Logging and utilities
python-dotenv>=1.0.1

# Type hints
typing-extensions>=4.0.0

# Pydantic for data validation
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
ruff>=0.6.1 